requirements.txt
psycopg[binary]
python-dotenv
streamlit>=1.25.0
